Incremental commit.  Compiles and runs.
Refactored & renamed some variables for clarity.  Resolved 2 (maybe 3?) bugs.
Populations were being initialized to entirely I1, rather than mostly S, because
the susceptible fraction was being cast as an int before multiplication.

Turned on movement between pops, and reintroductions during the burn-in, and
added a flag to prevent the simulation from repeatedly trying new simulations
if no paralytic cases are being observed.

Added some useful output to cerr, so we can tell what is going on with the sims.

Also, there were (perhaps rare) situations where we were incorrectly observing
an extinction: if a movement event happened between villages A and B, the states
of the individuals moving are I1 and IR or IR and I1, respectively, and those
people are the only infected people in those villages, village A would be logged
as having an extinction event.  Village B was correctly handled as not having
had an extinction event.  This is because extinction was evaluated for A before
the reciprocal movement had occurred.

I feel like there was one more bug, but I can't remember what it was right now.