Bugfix affecting movement between villages
When modeling movement, we were allowing the "other" village to be the same as v,
nullifying the movement event. This meant that if there were fewer villages,
more movement events were thrown out, effectively making the movement rate
dependent on number of villages in a rather unintuitive way.  Similarly, if a
population was subdivided into more villages, the total amount of movement
increased, making it less likely small populations would have independent
dynamics.