Modifying changes in commit 3c31fc1
A recent commit noted a bug in the movement model.  In retrospect, it's
less clear what the "right" model is, so this commit adds the ability to
choose the old movement model (not STRICT_TRAVEL) or the new one
(STRICT_TRAVEL), where the relative size of the patch does not affect
movement probability.

These two movement approaches mirror the concepts behind density-dependent
(old/not strict model) and frequency-dependent (new/strict model) transmission.