Changing screen and file outputing system
Revised the information that is printed to the screen to include all
villages.

Added initial support for new log system to save inital post-burn-in
states and observed event counts to a file.

Changed obs_time checks to be >= 0 instead of > 0 to make sure the
first observed event is included in logging.

Reordered the main event loop to ensure proper sequence of updating
times, sampling and handling events, and logging data.