Changes to output file system
Can control wheter event counts are logged using par->output_event_counts.

If this is false, the file will contain a decimal representation of a
three-digit binary number. Each digit represents whether that type of
event occured (regardless of count) on that day. For example, 7 in the
file represents "111" which means for that day, there were reported events
for AFP, ES, and EXT (in that order).

Also removed the checksum from the file (it will only be part of the
filename) and changed the replicate to hex (both to reduce file size).